+++ Assets/Scripts/Character/PartyBase.cs
@@ -56,7 +56,7 @@ public class PartyBase : FighterBase
         }
     }
 
-    public override void Damage(float damageAmount)
+    public override void Damage(int damageAmount)
     {
         // call base class
         base.Damage(damageAmount);
